requirements.txt
asgiref==3.3.1
autopep8 @ file:///home/conda/feedstock_root/build_artifacts/autopep8_1615918605177/work
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.7
django-filter==2.4.0
djangorestframework==3.12.2
idna==2.10
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
openapi-codec==1.3.2
psycopg2-binary==2.8.6
pycodestyle @ file:///home/conda/feedstock_root/build_artifacts/pycodestyle_1615833610040/work
pytz==2021.1
requests==2.25.1
simplejson==3.17.2
sqlparse==0.4.1
toml @ file:///home/conda/feedstock_root/build_artifacts/toml_1604308577558/work
uritemplate==3.0.1
urllib3==1.26.4
